requirements.txt
streamlit
google-generativeai
python-dotenv
pypdf
beautifulsoup4
requests
plotly
pandas
